third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEConvolveMatrixElement-dom-edgeMode-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'edgeMode' attribute of the SVGFEConvolveMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS convolveMatrixElement.getAttribute('edgeMode') is "wrap"
PASS convolveMatrixElement.getAttribute('edgeMode') is "duplicate"
PASS successfullyParsed is true

TEST COMPLETE
